// src/app/model/CasinoItem.ts
export class CasinoItems {
  id?: number;

  BonusStatus?: boolean;

  Name?: string;

  LocationName?: string;

  WebName?: string;

  SupportName?: string;

  VpnStatus?: boolean;

  GameCount?: number;

  PaymentCount?: number;

  LicenseCount?: number;

  StatusPoint?: boolean;

  Type?: number;

  CheckStatusLocation?: boolean;

  CheckStatusWeb?: boolean;

  CheckStatusCustomer?: boolean;

  CheckStatusLivechat?: boolean;

  PerfectPoint: any;

  IconsLocation?: number;

  ImageLocationPath?: string;
}